+++ package.json
@@ -31,13 +31,13 @@
     "packages/common",
     "packages/ui",
     "packages/core",
-    "packages/scaffold",
     "packages/scaffold-utils",
+    "packages/siwe",
+    "packages/scaffold",
     "packages/scaffold-react",
     "packages/scaffold-vue",
     "packages/polyfills",
     "packages/wallet",
-    "packages/siwe",
     "packages/wagmi",
     "packages/ethers",
     "packages/ethers5",
