+++ packages/core/index.ts
@@ -26,12 +26,6 @@ export type {
   ConnectionControllerState
 } from './src/controllers/ConnectionController.js'
 
-export { SIWEController } from './src/controllers/SIWEController.js'
-export type {
-  SIWEControllerClient,
-  SIWEControllerClientState
-} from './src/controllers/SIWEController.js'
-
 export { ConnectorController } from './src/controllers/ConnectorController.js'
 export type { ConnectorControllerState } from './src/controllers/ConnectorController.js'
