+++ packages/core/tsconfig.json
@@ -5,5 +5,5 @@
     "declarationDir": "./dist/types"
   },
   "extends": "../../tsconfig.json",
-  "include": ["index.ts", "src", "tests"]
+  "include": ["index.ts", "src", "tests", "../siwe/core/SIWEController.ts"]
 }
