+++ packages/scaffold-utils/src/EthersHelpersUtil.ts
@@ -1,4 +1,4 @@
-import type { CaipNetwork } from '@web3modal/scaffold'
+import type { CaipNetwork } from '@web3modal/core'
 import { ConstantsUtil } from './ConstantsUtil.js'
 import { PresetsUtil } from './PresetsUtil.js'
 import type { Chain, Provider } from './EthersTypesUtil.js'
