+++ packages/scaffold/index.ts
@@ -12,7 +12,6 @@ export * from './src/views/w3m-all-wallets-view/index.js'
 export * from './src/views/w3m-buy-in-progress-view/index.js'
 export * from './src/views/w3m-connect-view/index.js'
 export * from './src/views/w3m-connecting-external-view/index.js'
-export * from './src/views/w3m-connecting-siwe-view/index.js'
 export * from './src/views/w3m-connecting-wc-view/index.js'
 export * from './src/views/w3m-downloads-view/index.js'
 export * from './src/views/w3m-get-wallet-view/index.js'
@@ -42,7 +41,6 @@ export * from './src/partials/w3m-connecting-wc-browser/index.js'
 export * from './src/partials/w3m-connecting-wc-desktop/index.js'
 export * from './src/partials/w3m-connecting-wc-mobile/index.js'
 export * from './src/partials/w3m-connecting-wc-qrcode/index.js'
-export * from './src/partials/w3m-connecting-siwe/index.js'
 export * from './src/partials/w3m-connecting-wc-unsupported/index.js'
 export * from './src/partials/w3m-connecting-wc-web/index.js'
 export * from './src/partials/w3m-header/index.js'
