+++ packages/siwe/core/utils/ConstantsUtil.ts
