+++ packages/siwe/exports/index.ts
@@ -1,8 +1,25 @@
-import type { SIWEConfig, SIWESession } from '@web3modal/core'
+import type {
+  SIWEConfig,
+  SIWESession,
+  SIWECreateMessageArgs,
+  SIWEVerifyMessageArgs,
+  SIWEClientMethods
+} from '../core/utils/TypeUtils.js'
 import { Web3ModalSIWEClient } from '../src/client.js'
+export { SIWEController } from '../core/controller/SIWEController.js'
 
-export type { Web3ModalSIWEClient, SIWEConfig, SIWESession }
+export type {
+  Web3ModalSIWEClient,
+  SIWEConfig,
+  SIWESession,
+  SIWECreateMessageArgs,
+  SIWEVerifyMessageArgs,
+  SIWEClientMethods
+}
 
 export function createSIWEConfig(siweConfig: SIWEConfig) {
   return new Web3ModalSIWEClient(siweConfig)
 }
+
+export * from '../scaffold/partials/w3m-connecting-siwe/index.js'
+export * from '../scaffold/views/w3m-connecting-siwe-view/index.js'
