+++ packages/siwe/package.json
@@ -19,26 +19,13 @@
     "@web3modal/core": "4.0.7",
     "@web3modal/scaffold-utils": "4.0.7"
   },
+  "devDependencies": {
+    "valtio": "1.11.2",
+    "lit": "3.1.0"
+  },
   "peerDependencies": {
-    "react": ">=17",
-    "react-dom": ">=17",
-    "vue": ">=3"
-  },
-  "peerDependenciesMeta": {
-    "react": {
-      "optional": true
-    },
-    "react-dom": {
-      "optional": true
-    },
-    "vue": {
-      "optional": true
-    }
-  },
-  "optionalDependencies": {
-    "react": ">=17",
-    "react-dom": ">=17",
-    "vue": ">=3"
+    "valtio": ">=1 <2",
+    "lit": ">=3"
   },
   "keywords": [
     "web3",
