+++ packages/siwe/scaffold/partials/w3m-connecting-siwe/index.ts
