+++ packages/siwe/tsconfig.json
@@ -5,5 +5,5 @@
     "declarationDir": "./dist/types"
   },
   "extends": "../../tsconfig.json",
-  "include": ["exports", "src"]
+  "include": ["exports", "src", "core", "scaffold", "tests"]
 }
